move TaskFailedException to mrjob.sim
